requirements.txt
argon2-cffi==21.3.0
nava==0.7
tksheet==7.4.21
dotenv==1.1.0
pymongo==4.13.0